requirement.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
openai==1.12.0
pyyaml==6.0.1
python-dotenv==1.0.0